move DataTable to a separate file
